refactor: validator service layer에 위치하도록 수정(#520)
