+++ backend/src/main/java/com/minibeit/businessprofile/service/BusinessProfileService.java
@@ -1,7 +1,6 @@
 package com.minibeit.businessprofile.service;
 
 import com.minibeit.businessprofile.domain.BusinessProfile;
-import com.minibeit.businessprofile.domain.BusinessValidator;
 import com.minibeit.businessprofile.domain.UserBusinessProfile;
 import com.minibeit.businessprofile.domain.repository.BusinessProfileRepository;
 import com.minibeit.businessprofile.domain.repository.UserBusinessProfileRepository;
