+++ backend/src/main/java/com/minibeit/businessprofile/service/BusinessValidator.java
@@ -1,5 +1,7 @@
-package com.minibeit.businessprofile.domain;
+package com.minibeit.businessprofile.service;
 
+import com.minibeit.businessprofile.domain.BusinessProfile;
+import com.minibeit.businessprofile.domain.UserBusinessProfile;
 import com.minibeit.businessprofile.domain.repository.BusinessProfileRepository;
 import com.minibeit.common.exception.DuplicateException;
 import com.minibeit.common.exception.InvalidOperationException;
