+++ backend/src/main/java/com/minibeit/businessprofile/service/integrate/UserBusinessProfiles.java
@@ -1,0 +1,5 @@
+package com.minibeit.businessprofile.service.integrate;
+
+public interface UserBusinessProfiles {
+    Boolean existsByUserIdAndBusinessProfileId(Long businessProfileId, Long userId);
+}
