+++ backend/src/main/java/com/minibeit/user/service/UserService.java
@@ -5,7 +5,6 @@ import com.minibeit.file.service.integrate.Avatars;
 import com.minibeit.school.domain.School;
 import com.minibeit.school.service.integrate.Schools;
 import com.minibeit.user.domain.User;
-import com.minibeit.user.domain.UserValidator;
 import com.minibeit.user.domain.UserVerificationCode;
 import com.minibeit.user.domain.repository.UserRepository;
 import com.minibeit.user.domain.repository.UserVerificationCodeRepository;
